`Page` directory consolidates components: - Copyright - Page (root) - ProTip - RootContainer
