+++ apps/emergence/app/about/page.tsx
@@ -1,7 +1,7 @@
 import { type ReactElement } from 'react'
 import { Button as Link } from '@repo/mui/components/link/Button'
 import { Compact } from '@repo/mui/components/box/Compact'
-import { Page } from '@repo/mui/components/Page'
+import { Page } from '@repo/mui/components/Page/Page'
 
 export default function About(): ReactElement {
   return (
